refactor(server): move sandbox store helpers
