+++ scripts/max_lines_allowlist.txt
@@ -32,7 +32,7 @@ server/proliferate/db/store/cloud_mobility.py 1410 server structure migration de
 server/proliferate/db/store/cloud_repo_config.py 719 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_sync/commands.py 1565 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_sync/events.py 1109 server structure migration debt for oversized store file
-server/proliferate/db/store/cloud_workspaces.py 1482 server structure migration debt for oversized store file
+server/proliferate/db/store/cloud_workspaces.py 1187 server structure migration debt for oversized store file
 server/proliferate/server/billing/service.py 2459 server structure migration debt for oversized service file during billing orchestration consolidation
 server/proliferate/server/cloud/commands/domain/rules.py 607 server structure migration debt for oversized pure-domain file
 server/proliferate/server/cloud/mobility/service.py 1208 server structure migration debt for oversized service file
@@ -47,7 +47,7 @@ server/tests/integration/test_billing_api.py 2052 existing server oversized inte
 server/tests/integration/test_cloud_agent_auth_api.py 1664 existing server oversized integration test
 server/tests/integration/test_cloud_api.py 2346 existing server oversized integration test
 server/tests/integration/test_cloud_commands_api.py 4890 existing server oversized integration test
-server/tests/integration/test_sandbox_profile_foundation.py 1224 existing server oversized integration test
+server/tests/integration/test_sandbox_profile_foundation.py 1223 existing server oversized integration test
 server/tests/integration/test_schema_migrations.py 925 existing server oversized integration test
 server/tests/integration/test_stripe_webhooks.py 1456 existing server oversized integration test
 server/tests/unit/test_cloud_mobility_service.py 806 existing server oversized unit test
