+++ server/proliferate/server/cloud/runtime/provision.py
@@ -23,9 +23,7 @@ from proliferate.constants.cloud import (
     CloudTargetStatus,
     CloudWorkspaceStatus,
 )
-from proliferate.constants.cloud import (
-    WorkspaceStatus as LegacyWorkspaceStatus,
-)
+from proliferate.constants.cloud import WorkspaceStatus as LegacyWorkspaceStatus
 from proliferate.db import engine as db_engine
 from proliferate.db.store import cloud_sandboxes
 from proliferate.db.store.cloud_agent_auth import store as agent_auth_store
@@ -39,15 +37,17 @@ from proliferate.db.store.cloud_runtime_environments import (
     load_runtime_environment_with_sandbox,
     save_runtime_environment_state,
 )
-from proliferate.db.store.cloud_sync import targets as targets_store
-from proliferate.db.store.cloud_workspaces import (
+from proliferate.db.store.cloud_sandboxes import (
     bind_allocated_sandbox,
-    finalize_workspace_provision_for_ids,
     load_cloud_sandbox_by_id,
-    load_cloud_workspace_by_id,
-    mark_workspace_error_by_id,
     save_sandbox_provider_state,
     update_sandbox_status,
+)
+from proliferate.db.store.cloud_sync import targets as targets_store
+from proliferate.db.store.cloud_workspaces import (
+    finalize_workspace_provision_for_ids,
+    load_cloud_workspace_by_id,
+    mark_workspace_error_by_id,
     update_workspace_status_by_id,
 )
 from proliferate.db.store.users import load_user_with_oauth_accounts_by_id
